+++ moveit_core/kinematic_constraints/__init__.py
@@ -1,0 +1,1 @@
+from pymoveit_core.kinematic_constraints import *
