+++ moveit_core/setup.py
@@ -5,6 +5,7 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 d = generate_distutils_setup(
     packages=['moveit_core'],
+    package_dir={'': 'src'}
 )
 
 setup(**d)
