+++ moveit_core/src/moveit_core/collision_detection/__init__.py
@@ -1,0 +1,1 @@
+from pymoveit_core.collision_detection import *
