+++ moveit_core/src/moveit_core/planning_scene/__init__.py
@@ -1,0 +1,1 @@
+from pymoveit_core.planning_scene import *
