+++ moveit_core/src/moveit_core/robot_model/__init__.py
@@ -1,0 +1,1 @@
+from pymoveit_core.robot_model import *
